Function dissapointment_island::config::load

pub fn load(path: &str) -> Config

Load configuration data from a path. returns a Config struct.

This function expects to be passed in a valid YAML file that has YAML for each attribute in Config.

path - Path to desired YAML file.

What data should be in your configuration file

width - Width of the window in characters.

height - Height of the window in characters.

fullscreen - Determines whether or not game will start in fullscreen mode.

fontpath - Path to desired font to use.

fonttype - Type of font, either Default or Greyscale.

fontlayout - Layout of font. Either Tcod, AsciiInRow, or AsciiInCol.

renderer - Desired renderer to use. Either SDL, GLSL, or OpenGL.

Panics

This function will panic if:

The path is invalid

The file is invalid

The file is not a YAML file

The file is missing YAML for any attribute of Config

The YAML for each attribute is not the correct type

The YAML for fonttype, fontlayout, or renderer are not in their tcod enums

This is definitely a very touchy function but it is important that there are no errors with the configuration file because initializing the root console depends heavily on it.